Enhancement: add post list tests and clean up test folders
