+++ package.json
@@ -20,7 +20,7 @@
     "@types/node": "^16.18.16",
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
-    "axios": "^1.3.5",
+    "axios": "^1.4.0",
     "classnames": "^2.3.2",
     "date-fns": "^2.29.3",
     "react": "^18.2.0",
@@ -63,7 +63,7 @@
   "devDependencies": {
     "@types/testing-library__jest-dom": "^5.14.5",
     "@types/uuid": "^9.0.1",
-    "msw": "^1.2.1",
+    "msw": "^1.2.2",
     "prettier": "^2.8.7",
     "react-dropzone": "^14.2.3"
   },
